Return corret attrs from process_attrs method
If this doesn't return, it will cause problems when
`based_near` aka `location` is blank in the CSV.
